refactor(tests): extract search_symbols.rs (1 test, standalone)
Move search_symbols_uses_named_realm from mod.rs into
engine/tests/search_symbols.rs.

Open-at-execution decision: standalone file rather than absorbed into
engine::search::tests. The MCP-level test exercises realm routing through
full RuntimeEngine dispatch; unit-level search ranking tests live in
engine::search::tests (inside src/engine/search.rs). Different concerns.

Tests: 135 passed / 0 failed / 1 ignored (baseline unchanged).

After this commit, mod.rs contains only shared fixtures and module
declarations -- no inlined tests remain.

marky-n1h step 8 (5/5).
