+++ markymark-mcp/src/engine/tests/mod.rs
@@ -39,43 +39,8 @@ mod hash_embedding;
 mod outline;
 mod recommend;
 mod rename;
+mod search_symbols;
 mod workspace_scan;
 
 #[cfg(feature = "semantic-search")]
 mod preview_profiling;
-
-#[tokio::test]
-async fn search_symbols_uses_named_realm() {
-    let dir = make_temp_realm_dir();
-    fs::write(dir.path().join("doc.md"), "# UniqueHeadingXYZ\n").unwrap();
-    let engine = make_engine_with_custom_realm("search-realm", dir.path()).await;
-
-    // Default realm should return no matches for the unique heading
-    let result = engine
-        .execute(CoreOperation::SearchSymbols {
-            query: "UniqueHeadingXYZ".to_string(),
-            realm: None,
-        })
-        .await;
-    if let CoreOperationResult::Symbols(matches) = result {
-        assert!(
-            matches.is_empty(),
-            "default realm should not have the heading"
-        );
-    } else {
-        panic!("expected Symbols result");
-    }
-
-    // Named realm should find it
-    let result = engine
-        .execute(CoreOperation::SearchSymbols {
-            query: "UniqueHeadingXYZ".to_string(),
-            realm: Some("search-realm".to_string()),
-        })
-        .await;
-    if let CoreOperationResult::Symbols(matches) = result {
-        assert!(!matches.is_empty(), "named realm should have the heading");
-    } else {
-        panic!("expected Symbols result");
-    }
-}
